+++ SeriesGuide/src/com/battlelancer/seriesguide/util/ShareUtils.java
@@ -8,6 +8,7 @@ import com.battlelancer.seriesguide.getglueapi.PrepareRequestTokenActivity;
 import com.battlelancer.seriesguide.provider.SeriesContract.Episodes;
 import com.battlelancer.seriesguide.ui.SeriesGuidePreferences;
 import com.battlelancer.seriesguide.ui.ShowInfoActivity;
+import com.battlelancer.seriesguide.util.ShareUtils.TraktTask.OnTraktActionCompleteListener;
 import com.jakewharton.apibuilder.ApiException;
 import com.jakewharton.trakt.ServiceManager;
 import com.jakewharton.trakt.TraktException;
@@ -16,6 +17,7 @@ import com.jakewharton.trakt.enumerations.Rating;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
+import android.content.ContentValues;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.DialogInterface.OnClickListener;
@@ -89,9 +91,11 @@ public class ShareUtils {
      * @param shareData - a {@link Bundle} including all
      *            {@link ShareUtils.ShareItems}
      * @param shareMethod the {@link ShareMethod} to use
+     * @param listener listener will be notified once trakt task is finished
+     *            (only first time)
      */
     public static void onShareEpisode(FragmentActivity activity, Bundle shareData,
-            ShareMethod shareMethod) {
+            ShareMethod shareMethod, OnTraktActionCompleteListener listener) {
         final FragmentManager fm = activity.getSupportFragmentManager();
         final String imdbId = shareData.getString(ShareUtils.ShareItems.IMDBID);
         final String sharestring = shareData.getString(ShareUtils.ShareItems.SHARESTRING);
@@ -140,7 +144,7 @@ public class ShareUtils {
             case MARKSEEN_TRAKT: {
                 // trakt mark as seen
                 shareData.putInt(ShareItems.TRAKTACTION, TraktAction.SEEN_EPISODE.index());
-                new TraktTask(activity, fm, shareData).execute();
+                new TraktTask(activity, fm, shareData, listener).execute();
                 break;
             }
             case RATE_TRAKT: {
@@ -345,6 +349,12 @@ public class ShareUtils {
 
         private TraktAction mAction;
 
+        private OnTraktActionCompleteListener mListener;
+
+        public interface OnTraktActionCompleteListener {
+            public void onTraktActionComplete();
+        }
+
         /**
          * Do the specified TraktAction. traktData should include all required
          * parameters.
@@ -358,6 +368,12 @@ public class ShareUtils {
             mContext = context;
             mFm = manager;
             mTraktData = traktData;
+        }
+
+        public TraktTask(Context context, FragmentManager manager, Bundle traktData,
+                OnTraktActionCompleteListener listener) {
+            this(context, manager, traktData);
+            mListener = listener;
         }
 
         @Override
@@ -403,6 +419,17 @@ public class ShareUtils {
                         r = new Response();
                         r.status = TraktStatus.SUCCESS;
                         r.message = mContext.getString(R.string.trakt_seen);
+
+                        // if successful mark episode as seen locally
+                        // immediately
+                        ContentValues values = new ContentValues();
+                        values.put(Episodes.WATCHED, true);
+                        mContext.getContentResolver().update(
+                                Episodes.buildEpisodesOfShowUri(String.valueOf(tvdbid)), values,
+                                Episodes.SEASON + "=? AND " + Episodes.NUMBER + "=?", new String[] {
+                                        String.valueOf(season), String.valueOf(episode)
+                                });
+                        mContext.getContentResolver().notifyChange(Episodes.CONTENT_URI, null);
                         break;
                     }
                     case RATE_EPISODE: {
@@ -472,6 +499,10 @@ public class ShareUtils {
                 TraktCredentialsDialogFragment newFragment = TraktCredentialsDialogFragment
                         .newInstance(mTraktData);
                 newFragment.show(ft, "traktdialog");
+            }
+
+            if (mListener != null) {
+                mListener.onTraktActionComplete();
             }
         }
     }